+++ docs/changelog.md
@@ -3,3 +3,9 @@
 - The game now automatically uses your username for your save name. You can still use custom save names by passing '-s' when running the program.
 - You can now pass the '-d' argument with a value of 1-3 (inclusive) to bypass difficulty selection.
 - You can also pass '-h', '--help', or '--usage' to get a help document, or '-v' for version information.
+
+## v1.2
+- Reading from a score file now has it's own arguments, '-f' and '--readfile'
+- Score files now use the .scf extension, which is now required for the game to read from the file.
+- Fixed the '--usage' arg not being handled at all even when passed.
+- '-v' now has a corresponding long option '--version'